+++ .har/factory-line/cases/next-pages-api/files/pages/api/user/[id].ts
@@ -1,0 +1,19 @@
+import type { NextApiRequest, NextApiResponse } from 'next'
+
+export default function userHandler(req: NextApiRequest, res: NextApiResponse) {
+  const { query, method } = req
+  const id = parseInt(query.id as string, 10)
+  const name = query.name as string
+
+  switch (method) {
+    case 'GET':
+      res.status(200).json({ id, name: `User ${id}` })
+      break
+    case 'PUT':
+      res.status(200).json({ id, name: name || `User ${id}` })
+      break
+    default:
+      res.setHeader('Allow', ['GET', 'PUT'])
+      res.status(405).end(`Method ${method} Not Allowed`)
+  }
+}